Confirm page
when the user clicks the "Confirm Reset" button, he will be redirected depending on the result of the confirmReset function. If the password reset request is successful, the user is redirected to: /login/confirm page. If there is an error during the password reset request, the user will remain on the forgot password page /login/forgot-password